requirements.txt
chromadb==0.4.3
ibm-cloud-sdk-core==3.16.7
ibm-watson-machine-learning==1.0.342
langchain>=0.0.318,<0.1.0
plotly==5.15.0
PyMuPDF==1.23.2
pypdf==3.15.4
python-dotenv==1.0.0
sentence-transformers==2.2.2
streamlit==1.24.1
beautifulsoup4>=4.12.2
spacy>=2.2.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm